LayoutTests/inspector/dom-debugger/url-breakpoints-all-requests-expected.txt

CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 1
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 2
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 3
CONSOLE MESSAGE: BREAKPOINT ACTION LOG 4
Tests for All Requests URL breakpoint.


== Running test suite: URLBreakpoint.BreakOnAll
-- Running test case: URLBreakpoint.BreakOnAll.Fetch
Adding All Requests URL Breakpoint...
Request Fetch "resources/dataFetch.json"...
PAUSED:
PASS: Pause reason should be Fetch.
PASS: Pause data URL should be empty.
CALL STACK:
0: [F] loadResourceFetch
1: [P] Global Code
-- Running test teardown.

-- Running test case: URLBreakpoint.BreakOnAll.Fetch.Options.Condition
Adding All Requests URL Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnAll.Fetch.Options.IgnoreCount
Adding All Requests URL Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should not pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnAll.Fetch.Options.Action.Log
Adding All Requests URL Breakpoint...

Adding log action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnAll.Fetch.Options.Actions.Evaluate
Adding All Requests URL Breakpoint...

Adding evaluate action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
Request Fetch "resources/dataFetch.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnAll.XHR
Adding All Requests URL Breakpoint...
Request XHR "resources/dataXHR.json"...
PAUSED:
PASS: Pause reason should be XHR.
PASS: Pause data URL should be empty.
CALL STACK:
0: [F] loadResourceXHR
1: [P] Global Code
-- Running test teardown.

-- Running test case: URLBreakpoint.BreakOnAll.XHR.Options.Condition
Adding All Requests URL Breakpoint...

Setting condition to 'false'...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Setting condition to 'true'...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnAll.XHR.Options.IgnoreCount
Adding All Requests URL Breakpoint...

Setting ignoreCount to '2'...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should not pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should pause.

-- Running test case: URLBreakpoint.BreakOnAll.XHR.Options.Action.Log
Adding All Requests URL Breakpoint...

Adding log action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing log action...
Enabling auto-continue...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing log action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

-- Running test case: URLBreakpoint.BreakOnAll.XHR.Options.Actions.Evaluate
Adding All Requests URL Breakpoint...

Adding evaluate action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should pause.

Editing evaluate action...
Enabling auto-continue...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.

Editing evaluate action...

Triggering breakpoint...
Request XHR "resources/dataXHR.json"...
PASS: Should execute breakpoint action.
PASS: Should not pause.
